LayoutTests/platform/glib/imported/w3c/web-platform-tests/WebCryptoAPI/derive_bits_keys/cfrg_curves_bits.https.any.worker-expected.txt


PASS setup - define tests
PASS X25519 good parameters
PASS X25519 mixed case parameters
PASS X25519 with null length
PASS X25519 short result
PASS X25519 non-multiple of 8 bits
PASS X25519 missing public property
PASS X25519 public property of algorithm is not a CryptoKey
FAIL X25519 mismatched algorithms assert_equals: Should throw correct error, not TypeError: Type error expected "InvalidAccessError" but got "TypeError"
PASS X25519 no deriveBits usage for base key
PASS X25519 base key is not a private key
PASS X25519 public property value is a private key
PASS X25519 public property value is a secret key
PASS X25519 asking for too many bits
FAIL X448 good parameters assert_unreached: deriveBits failed with error TypeError: Argument 2 ('baseKey') to SubtleCrypto.deriveBits must be an instance of CryptoKey Reached unreachable code
FAIL X448 mixed case parameters assert_unreached: deriveBits failed with error TypeError: Argument 2 ('baseKey') to SubtleCrypto.deriveBits must be an instance of CryptoKey Reached unreachable code
FAIL X448 with null length assert_unreached: deriveBits failed with error TypeError: Argument 2 ('baseKey') to SubtleCrypto.deriveBits must be an instance of CryptoKey Reached unreachable code
FAIL X448 short result assert_unreached: deriveBits failed with error TypeError: Argument 2 ('baseKey') to SubtleCrypto.deriveBits must be an instance of CryptoKey Reached unreachable code
FAIL X448 non-multiple of 8 bits assert_unreached: deriveBits failed with error TypeError: Argument 2 ('baseKey') to SubtleCrypto.deriveBits must be an instance of CryptoKey Reached unreachable code
PASS X448 missing public property
PASS X448 public property of algorithm is not a CryptoKey
FAIL X448 mismatched algorithms assert_equals: Should throw correct error, not TypeError: Argument 2 ('baseKey') to SubtleCrypto.deriveBits must be an instance of CryptoKey expected "InvalidAccessError" but got "TypeError"
FAIL X448 no deriveBits usage for base key assert_equals: Should throw correct error, not TypeError: Argument 2 ('baseKey') to SubtleCrypto.deriveBits must be an instance of CryptoKey expected "InvalidAccessError" but got "TypeError"
FAIL X448 base key is not a private key assert_equals: Should throw correct error, not TypeError: Argument 2 ('baseKey') to SubtleCrypto.deriveBits must be an instance of CryptoKey expected "InvalidAccessError" but got "TypeError"
FAIL X448 public property value is a private key assert_equals: Should throw correct error, not TypeError: Argument 2 ('baseKey') to SubtleCrypto.deriveBits must be an instance of CryptoKey expected "InvalidAccessError" but got "TypeError"
FAIL X448 public property value is a secret key assert_equals: Should throw correct error, not TypeError: Argument 2 ('baseKey') to SubtleCrypto.deriveBits must be an instance of CryptoKey expected "InvalidAccessError" but got "TypeError"
FAIL X448 asking for too many bits assert_equals: Should throw correct error, not TypeError: Argument 2 ('baseKey') to SubtleCrypto.deriveBits must be an instance of CryptoKey expected "OperationError" but got "TypeError"
